+++ Builds/CMake/deps/Protobuf.cmake
@@ -129,27 +129,28 @@ else ()
   endif ()
 endif ()
 
-file (MAKE_DIRECTORY ${CMAKE_BINARY_DIR}/proto_gen)
-set (save_CBD ${CMAKE_CURRENT_BINARY_DIR})
-set (CMAKE_CURRENT_BINARY_DIR ${CMAKE_BINARY_DIR}/proto_gen)
-protobuf_generate_cpp (
-  PROTO_SRCS
-  PROTO_HDRS
-  src/ripple/proto/ripple.proto)
-set (CMAKE_CURRENT_BINARY_DIR ${save_CBD})
+set(output_dir ${CMAKE_BINARY_DIR}/proto_gen)
+file(MAKE_DIRECTORY ${output_dir})
+set(ccbd ${CMAKE_CURRENT_BINARY_DIR})
+set(CMAKE_CURRENT_BINARY_DIR ${output_dir})
+protobuf_generate_cpp(PROTO_SRCS PROTO_HDRS src/ripple/proto/ripple.proto)
+set(CMAKE_CURRENT_BINARY_DIR ${ccbd})
 
-add_library (pbufs STATIC ${PROTO_SRCS} ${PROTO_HDRS})
-
-target_include_directories (pbufs PRIVATE src)
-target_include_directories (pbufs
-  SYSTEM PUBLIC ${CMAKE_BINARY_DIR}/proto_gen)
-target_link_libraries (pbufs protobuf::libprotobuf)
-target_compile_options (pbufs
+target_include_directories(xrpl_core SYSTEM PUBLIC
+    # The generated implementation imports the header relative to the output
+    # directory.
+    $<BUILD_INTERFACE:${output_dir}>
+    $<BUILD_INTERFACE:${output_dir}/src>
+)
+target_sources(xrpl_core PRIVATE ${output_dir}/src/ripple/proto/ripple.pb.cc)
+install(
+  FILES ${output_dir}/src/ripple/proto/ripple.pb.h
+  DESTINATION include/ripple/proto)
+target_link_libraries(xrpl_core PUBLIC protobuf::libprotobuf)
+target_compile_options(xrpl_core
   PUBLIC
     $<$<BOOL:${is_xcode}>:
       --system-header-prefix="google/protobuf"
       -Wno-deprecated-dynamic-exception-spec
-    >)
-add_library (Ripple::pbufs ALIAS pbufs)
-target_link_libraries (ripple_libs INTERFACE Ripple::pbufs)
-exclude_if_included (pbufs)
+    >
+)
