+++ src/ripple/app/ledger/InboundLedger.h
@@ -197,14 +197,6 @@ private:
     std::unique_ptr<PeerSet> mPeerSet;
 };
 
-/** Deserialize a ledger header from a byte array. */
-LedgerInfo
-deserializeHeader(Slice data, bool hasHash = false);
-
-/** Deserialize a ledger header (prefixed with 4 bytes) from a byte array. */
-LedgerInfo
-deserializePrefixedHeader(Slice data, bool hasHash = false);
-
 }  // namespace ripple
 
 #endif
