+++ src/ripple/app/ledger/impl/InboundLedger.cpp
@@ -269,36 +269,6 @@ InboundLedger::neededStateHashes(int max, SHAMapSyncFilter* filter) const
         mLedger->info().accountHash, mLedger->stateMap(), max, filter);
 }
 
-LedgerInfo
-deserializeHeader(Slice data, bool hasHash)
-{
-    SerialIter sit(data.data(), data.size());
-
-    LedgerInfo info;
-
-    info.seq = sit.get32();
-    info.drops = sit.get64();
-    info.parentHash = sit.get256();
-    info.txHash = sit.get256();
-    info.accountHash = sit.get256();
-    info.parentCloseTime =
-        NetClock::time_point{NetClock::duration{sit.get32()}};
-    info.closeTime = NetClock::time_point{NetClock::duration{sit.get32()}};
-    info.closeTimeResolution = NetClock::duration{sit.get8()};
-    info.closeFlags = sit.get8();
-
-    if (hasHash)
-        info.hash = sit.get256();
-
-    return info;
-}
-
-LedgerInfo
-deserializePrefixedHeader(Slice data, bool hasHash)
-{
-    return deserializeHeader(data + 4, hasHash);
-}
-
 // See how much of the ledger data is stored locally
 // Data found in a fetch pack will be stored
 void
@@ -562,10 +532,9 @@ InboundLedger::trigger(std::shared_ptr<Peer> const& peer, TriggerReason reason)
 
     if (auto stream = journal_.debug())
     {
+        stream << "Trigger acquiring ledger " << hash_;
         if (peer)
-            stream << "Trigger acquiring ledger " << hash_ << " from " << peer;
-        else
-            stream << "Trigger acquiring ledger " << hash_;
+            stream << " from " << peer;
 
         if (complete_ || failed_)
             stream << "complete=" << complete_ << " failed=" << failed_;
