+++ src/ripple/app/reporting/P2pProxy.h
@@ -21,12 +21,11 @@
 #define RIPPLE_APP_REPORTING_P2PPROXY_H_INCLUDED
 
 #include <ripple/app/main/Application.h>
+#include <ripple/proto/org/xrpl/rpc/v1/xrp_ledger.grpc.pb.h>
 #include <ripple/rpc/Context.h>
 #include <ripple/rpc/impl/Handler.h>
 
 #include <boost/beast/websocket.hpp>
-
-#include "org/xrpl/rpc/v1/xrp_ledger.grpc.pb.h"
 #include <grpcpp/grpcpp.h>
 
 namespace ripple {
