+++ src/ripple/ledger/impl/View.cpp
@@ -158,23 +158,6 @@ cdirNext(
 //
 //------------------------------------------------------------------------------
 
-void
-addRaw(LedgerInfo const& info, Serializer& s, bool includeHash)
-{
-    s.add32(info.seq);
-    s.add64(info.drops.drops());
-    s.addBitString(info.parentHash);
-    s.addBitString(info.txHash);
-    s.addBitString(info.accountHash);
-    s.add32(info.parentCloseTime.time_since_epoch().count());
-    s.add32(info.closeTime.time_since_epoch().count());
-    s.add8(info.closeTimeResolution.count());
-    s.add8(info.closeFlags);
-
-    if (includeHash)
-        s.addBitString(info.hash);
-}
-
 bool
 hasExpired(ReadView const& view, std::optional<std::uint32_t> const& exp)
 {
