+++ src/ripple/overlay/impl/ProtocolMessage.h
@@ -32,7 +32,6 @@
 #include <cstdint>
 #include <memory>
 #include <optional>
-#include <ripple.pb.h>
 #include <type_traits>
 #include <vector>
