+++ src/ripple/protocol/NFTokenOfferID.h
@@ -17,42 +17,41 @@
 */
 //==============================================================================
 
-#ifndef RIPPLE_RPC_NFTSYNTHETICSERIALIZER_H_INCLUDED
-#define RIPPLE_RPC_NFTSYNTHETICSERIALIZER_H_INCLUDED
+#ifndef RIPPLE_PROTOCOL_NFTOKENOFFERID_H_INCLUDED
+#define RIPPLE_PROTOCOL_NFTOKENOFFERID_H_INCLUDED
 
-#include <ripple/protocol/Protocol.h>
-#include <ripple/protocol/STBase.h>
+#include <ripple/basics/base_uint.h>
+#include <ripple/json/json_forwards.h>
+#include <ripple/protocol/STTx.h>
+#include <ripple/protocol/TxMeta.h>
 
-#include <functional>
 #include <memory>
-
-namespace Json {
-class Value;
-}
+#include <optional>
 
 namespace ripple {
 
-class TxMeta;
-class STTx;
-
-namespace RPC {
-
-struct JsonContext;
-
 /**
-   Adds common synthetic fields to transaction-related JSON responses
+   Add an `offer_id` field to the `meta` output parameter.
+   The field is only added to successful NFTokenCreateOffer transactions.
 
+   Helper functions are not static because they can be used by Clio.
    @{
  */
+bool
+canHaveNFTokenOfferID(
+    std::shared_ptr<STTx const> const& serializedTx,
+    TxMeta const& transactionMeta);
+
+std::optional<uint256>
+getOfferIDFromCreatedOffer(TxMeta const& transactionMeta);
+
 void
-insertNFTSyntheticInJson(
-    Json::Value&,
-    RPC::JsonContext const&,
-    std::shared_ptr<STTx const> const&,
-    TxMeta const&);
+insertNFTokenOfferID(
+    Json::Value& response,
+    std::shared_ptr<STTx const> const& transaction,
+    TxMeta const& transactionMeta);
 /** @} */
 
-}  // namespace RPC
 }  // namespace ripple
 
 #endif
