+++ src/ripple/protocol/impl/NFTokenID.cpp
@@ -17,21 +17,10 @@
 */
 //==============================================================================
 
-#include <ripple/rpc/NFTokenID.h>
-
-#include <ripple/app/ledger/LedgerMaster.h>
-#include <ripple/app/ledger/OpenLedger.h>
-#include <ripple/app/misc/Transaction.h>
-#include <ripple/ledger/View.h>
-#include <ripple/net/RPCErr.h>
-#include <ripple/protocol/AccountID.h>
-#include <ripple/protocol/Feature.h>
-#include <ripple/rpc/Context.h>
-#include <ripple/rpc/impl/RPCHelpers.h>
-#include <boost/algorithm/string/case_conv.hpp>
+#include <ripple/protocol/NFTokenID.h>
+#include <ripple/protocol/jss.h>
 
 namespace ripple {
-namespace RPC {
 
 bool
 canHaveNFTokenID(
@@ -198,5 +187,4 @@ insertNFTokenID(
     }
 }
 
-}  // namespace RPC
 }  // namespace ripple
